requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.11.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0
python-jose==3.3.0
bcrypt==4.0.1
email-validator==2.2.0
python-multipart==0.0.9
httpx==0.27.2
pytest==7.4.0
